refactor tower, projectile, enemy
Removed all fields from scripts, added serializefield to all properties. I tried it before to avoid unnecessary code, but couldn't make it.
I was trying to use classic [SerializeField] before properties, but it should be [field: SerializeField] so it can reach "field" of property and let us edit the value in the Unity inspector.